build(heroku): updating Procfile and scripts
Using `node dist/src/index.js` as the entryproint in the `Procfile` for
heroku. It seems heroku only leaves the dependencies intact
(and not devDependencies). Typescript is a dev dep so there is no
tsc. In fact, it's already been built by the deploy script so we can
just run our app through node.

Changing the scripts in `package.json` back to using `run-s` instead of
`npm run`