Updating the database to support better combinations
The `makes` field was using a Record<string, string> to represent
what could be made. However, there was an assumption that an item
combined with another item only makes 1 item. That is an incorrect
assumption and has been fixed by having the `makes` field utilize
an array of source + target objects. So for example, `arctic` can
be combined with `dog` to make `seal` AND `husky`. Without this
change only one of those can exist in the record (with `dog` being
the key).

This also required some code updates.

Finally ensuring that a .nvmrc file exists to use the correct node
version.